requirements.txt
fitz==0.0.1.dev2
opendal==0.45.11
pandas==2.2.3
Pillow==11.0.0
pyarrow==17.0.0
pytesseract==0.3.13
python_magic==0.4.27
python_pdfbox==0.1.8.1
tika==2.6.0
